# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UploadResultPreviewItem = {[key: string]: string};

export interface UploadResult {
  sessionId: string;
  filename: string;
  rowCount: number;
  columns: string[];
  preview: UploadResultPreviewItem[];
}

export interface CureInput {
  /** Column name in the raw file to cure */
  rawColumn: string;
  /** Column name in the master list to match against */
  masterColumn: string;
  /** Output column name for the standardized value */
  outputColumnName?: string;
}

export type MatchRowMatchType = typeof MatchRowMatchType[keyof typeof MatchRowMatchType];


export const MatchRowMatchType = {
  exact: 'exact',
  normalized: 'normalized',
  fuzzy: 'fuzzy',
  none: 'none',
} as const;

export type MatchRowStatus = typeof MatchRowStatus[keyof typeof MatchRowStatus];


export const MatchRowStatus = {
  approved: 'approved',
  corrected: 'corrected',
  ignored: 'ignored',
  pending: 'pending',
} as const;

export interface MatchRow {
  rowIndex: number;
  originalValue: string;
  /** @nullable */
  suggestedValue?: string | null;
  confidenceScore: number;
  matchType: MatchRowMatchType;
  status: MatchRowStatus;
  /** @nullable */
  correctedValue?: string | null;
}

export interface CureResult {
  sessionId: string;
  matches: MatchRow[];
  totalRows: number;
  rawColumn: string;
  masterColumn: string;
  outputColumnName: string;
}

export type DecisionItemStatus = typeof DecisionItemStatus[keyof typeof DecisionItemStatus];


export const DecisionItemStatus = {
  approved: 'approved',
  corrected: 'corrected',
  ignored: 'ignored',
} as const;

export interface DecisionItem {
  rowIndex: number;
  status: DecisionItemStatus;
  /** @nullable */
  correctedValue?: string | null;
}

export interface DecisionsInput {
  decisions: DecisionItem[];
}

export interface DecisionUpdateResult {
  sessionId: string;
  updatedCount: number;
}

export interface QualitySummary {
  totalRows: number;
  uniqueRawValues: number;
  exactMatches: number;
  normalizedMatches: number;
  fuzzyMatches: number;
  needsReview: number;
  noMatch: number;
  blankValues: number;
  duplicateValues: number;
  approved: number;
  corrected: number;
  ignored: number;
  pending: number;
}

export interface ErrorResponse {
  error: string;
}