# lib/api-zod/src/generated/types/cureInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CureInput {
  /** Column name in the raw file to cure */
  rawColumn: string;
  /** Column name in the master list to match against */
  masterColumn: string;
  /** Output column name for the standardized value */
  outputColumnName?: string;
}