# lib/api-zod/src/generated/types/cureResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchRow } from './matchRow';

export interface CureResult {
  sessionId: string;
  matches: MatchRow[];
  totalRows: number;
  rawColumn: string;
  masterColumn: string;
  outputColumnName: string;
}