# lib/api-zod/src/generated/types/decisionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionItemStatus } from './decisionItemStatus';

export interface DecisionItem {
  rowIndex: number;
  status: DecisionItemStatus;
  /** @nullable */
  correctedValue?: string | null;
}