# lib/api-zod/src/generated/types/decisionUpdateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DecisionUpdateResult {
  sessionId: string;
  updatedCount: number;
}