# lib/api-zod/src/generated/types/matchRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchRowMatchType } from './matchRowMatchType';
import type { MatchRowStatus } from './matchRowStatus';

export interface MatchRow {
  rowIndex: number;
  originalValue: string;
  /** @nullable */
  suggestedValue?: string | null;
  confidenceScore: number;
  matchType: MatchRowMatchType;
  status: MatchRowStatus;
  /** @nullable */
  correctedValue?: string | null;
}