# lib/api-zod/src/generated/types/qualitySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QualitySummary {
  totalRows: number;
  uniqueRawValues: number;
  exactMatches: number;
  normalizedMatches: number;
  fuzzyMatches: number;
  needsReview: number;
  noMatch: number;
  blankValues: number;
  duplicateValues: number;
  approved: number;
  corrected: number;
  ignored: number;
  pending: number;
}