# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Foundry API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadResultPreviewItem } from './uploadResultPreviewItem';

export interface UploadResult {
  sessionId: string;
  filename: string;
  rowCount: number;
  columns: string[];
  preview: UploadResultPreviewItem[];
}